(more) secure sessions
Previously the entire ssid was being sent in the clear. Now we
only set a part of the ssid in the user's cookie and derive the rest with
something like sha512(sid + other info). This should prevent simple session hijacking.
Temporarily commented out memcache stuff. Seems like an optimization that can
come later.